[front] 非开发环境下应用不能启动多个

+++ sparkle-cli/electron/main.ts
@@ -7,8 +7,10 @@ import IpcChannels from "../common/IpcChannels";
 import { AppCfgStore } from "./config/AppStore";
 import { destroyMainWindow, initMainWindow } from "./window/WindowManager.ts";
 
+let mainWin: BrowserWindow | null;
+
 function createWin() {
-  const win = new BrowserWindow({
+  mainWin = new BrowserWindow({
     width: 800,
     height: 600,
     show: false,
@@ -19,52 +21,67 @@ function createWin() {
       sandbox: false
     }
   });
-  initMainWindow(win);
+  initMainWindow(mainWin);
 
-  win.on("ready-to-show", () => {
-    win.show();
+  mainWin.on("ready-to-show", () => {
+    if (!mainWin) {
+      return;
+    }
+    mainWin.show();
     if (is.dev) {
-      win.webContents.openDevTools();
+      mainWin.webContents.openDevTools();
     }
 
     const rootDir = AppCfgStore.get(AppRootDirKey);
     if (!rootDir) {
-      win.webContents.send(IpcChannels.ToInitPage);
+      mainWin.webContents.send(IpcChannels.ToInitPage);
     }
   });
 
   if (is.dev && process.env.VITE_DEV_SERVER_URL) {
-    win.loadURL(process.env.VITE_DEV_SERVER_URL);
+    mainWin.loadURL(process.env.VITE_DEV_SERVER_URL);
   } else {
-    win.loadFile("build/renderer/index.html");
+    mainWin.loadFile("build/renderer/index.html");
   }
 }
 
-app.whenReady().then(() => {
-  // Set app user model id for windows
-  electronApp.setAppUserModelId("com.cream.sparkle");
+const gotTheLock = app.requestSingleInstanceLock();
+if (gotTheLock || is.dev) {
+  app.whenReady().then(() => {
+    // Set app user model id for windows
+    electronApp.setAppUserModelId("com.cream.sparkle");
 
-  // Default open or close DevTools by F12 in development
-  // and ignore CommandOrControl + R in production.
-  // see https://github.com/alex8088/electron-toolkit/tree/master/packages/utils
-  app.on("browser-window-created", (_, window) => {
-    optimizer.watchWindowShortcuts(window);
+    // Default open or close DevTools by F12 in development
+    // and ignore CommandOrControl + R in production.
+    // see https://github.com/alex8088/electron-toolkit/tree/master/packages/utils
+    app.on("browser-window-created", (_, window) => {
+      optimizer.watchWindowShortcuts(window);
+    });
+
+    createWin();
+    mainIpcSetup();
+    app.on("second-instance", () => {
+      if (mainWin) {
+        mainWin.restore();
+        mainWin.show();
+        mainWin.focus();
+      }
+    });
+    app.on("activate", function () {
+      // mac特定处理
+      if (BrowserWindow.getAllWindows().length === 0) createWin();
+    });
   });
 
-  createWin();
-  mainIpcSetup();
-  app.on("activate", function () {
-    // mac特定处理
-    if (BrowserWindow.getAllWindows().length === 0) createWin();
+  // mac特定处理
+  app.on("window-all-closed", () => {
+    if (process.platform !== "darwin") {
+      app.quit();
+      mainWin = null;
+      destroyMainWindow();
+    }
   });
-});
-
-// mac特定处理
-app.on("window-all-closed", () => {
-  if (process.platform !== "darwin") {
-    app.quit();
-    destroyMainWindow();
-  }
-});
-
-// 下面可以包括主进程特定代码的其余部分。也可以将它们放在单独的文件中，并在这里引入。
+  // 下面可以包括主进程特定代码的其余部分。也可以将它们放在单独的文件中，并在这里引入。
+} else {
+  app.quit();
+}
